rcar-gen3: core-image-weston: Fix gfx-only build error
"gfx-only" build occurs error because it doesn't contain omx-user-module.
This patch fixes it by checking image directory before copying files.

+++ meta-rcar-gen3/recipes-core/images/core-image-weston.bbappend
@@ -19,11 +19,15 @@ remove_gfx_mmp_files() {
     rm -rf ${DEPLOY_DIR_IMAGE}/gfx_mmp
     mkdir -p ${DEPLOY_DIR_IMAGE}/gfx_mmp
 
-    cp -rpf ${BASE_WORKDIR}/${MACHINE}-poky-linux/gles-user-module/1.0-r0/image/* \
-        -t ${DEPLOY_DIR_IMAGE}/gfx_mmp
+    if [[ -e ${BASE_WORKDIR}/${MACHINE}-poky-linux/gles-user-module/1.0-r0/image ]]; then
+        cp -rpf ${BASE_WORKDIR}/${MACHINE}-poky-linux/gles-user-module/1.0-r0/image/* \
+            -t ${DEPLOY_DIR_IMAGE}/gfx_mmp
+    fi
 
-    cp -rpf ${BASE_WORKDIR}/${MACHINE}-poky-linux/omx-user-module/1.0-r0/image/* \
-        -t ${DEPLOY_DIR_IMAGE}/gfx_mmp
+    if [[ -e ${BASE_WORKDIR}/${MACHINE}-poky-linux/omx-user-module/1.0-r0/image ]]; then
+        cp -rpf ${BASE_WORKDIR}/${MACHINE}-poky-linux/omx-user-module/1.0-r0/image/* \
+            -t ${DEPLOY_DIR_IMAGE}/gfx_mmp
+    fi
 
     tar -jcf ${DEPLOY_DIR_IMAGE}/gfx_mmp.tar.bz2 -C ${DEPLOY_DIR_IMAGE}/gfx_mmp .
 }
